+++ src/users/users.service.ts
@@ -1,9 +1,4 @@
-import {
-  ConflictException,
-  Injectable,
-  NotFoundException,
-} from '@nestjs/common';
-import { CreateUserDto } from './dto/create-user.dto';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { FilterUsersDto } from './dto/filter-users.dto';
 import { Prisma, User } from '@prisma/client';
@@ -13,20 +8,8 @@ import { PaginatedResponse } from 'src/common/types/return-type';
 @Injectable()
 export class UsersService {
   constructor(private prisma: PrismaService) {}
-  async create(createUserDto: CreateUserDto): Promise<User | undefined> {
-    try {
-      const user = await this.prisma.user.create({ data: createUserDto });
-      return user;
-    } catch (error) {
-      if (error.code === 'P2002') {
-        throw new ConflictException(
-          'User with this email or username already exists',
-        );
-      }
-    }
-  }
 
-  async findAll({
+  async getUsers({
     limit: take = 10,
     page = 1,
     search,
@@ -67,13 +50,13 @@ export class UsersService {
     };
   }
 
-  async findOne(id: string): Promise<User> {
+  async getUserById(id: string): Promise<User> {
     const user = await this.prisma.user.findUnique({ where: { id } });
     if (!user) throw new NotFoundException(`User #${id} not found`);
     return user;
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto): Promise<User> {
+  async updateUser(id: string, updateUserDto: UpdateUserDto): Promise<User> {
     const user = await this.prisma.user.findUnique({ where: { id } });
     if (!user) throw new NotFoundException(`User #${id} not found`);
 
@@ -83,16 +66,10 @@ export class UsersService {
     });
   }
 
-  async remove(
-    id: string,
-  ): Promise<{ statusCode: number; message: string } | undefined> {
-    try {
-      await this.prisma.user.delete({ where: { id } });
-      return { statusCode: 200, message: 'User deleted successfully' };
-    } catch (error) {
-      if (error.code === 'P2025') {
-        throw new NotFoundException();
-      }
-    }
+  async deleteAccount(id: string) {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+    if (!user) throw new NotFoundException(`User #${id} not found`);
+
+    return this.prisma.user.delete({ where: { id } });
   }
 }
